requirements.txt
gradio
whisper
torch
python-dotenv
datasets
ipywidgets
transformers
numba
openai-whisper